Final tweaks from deployment server.
There were some changes made to the visualization that were not
uploaded to github on both the deployment server and in the
development environment. These are the changes from the dev
server.

Note that this was actually uploaded on Nov 25 but the date was
set backwards to indicate that the changes were made back in
April.